+++ src/main/java/io/codeagainsthumanity/models/AllDecks.java
@@ -1,0 +1,16 @@
+package io.codeagainsthumanity.models;
+
+import java.lang.reflect.Array;
+import java.util.ArrayList;
+
+public class AllDecks {
+
+    ArrayList <Object> blackCards;
+    ArrayList <Object> whiteCards;
+
+    String text;
+    int pick;
+
+    
+
+}
