+++ src/main/resources/static/cards.json
